+++ src/UI/SignTransactionsModals/components/DeviceConfirmationScreen.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import {
+  useGetAccount,
   useGetAccountProvider,
   useGetSignTransactionsError,
   useSignTransactionsCommonData
@@ -13,6 +14,7 @@ import {
 
 export const DeviceConfirmationScreen = ({
   Screen,
+  GuardianScreen,
   verifyReceiverScam,
   className
 }: ExtraConfirmationScreenPropsType) => {
@@ -30,6 +32,7 @@ export const DeviceConfirmationScreen = ({
 
   const signTransactionsError = useGetSignTransactionsError();
   const { providerType } = useGetAccountProvider();
+  const { isGuarded } = useGetAccount();
 
   const handleClose = () => {
     onAbort(transactionsToSign?.sessionId);
@@ -39,6 +42,8 @@ export const DeviceConfirmationScreen = ({
 
   const signProps: SignPropsType = {
     handleClose,
+    GuardianScreen,
+    isGuarded,
     error: signError,
     sessionId: transactionsToSign?.sessionId,
     transactions: transactionsToSign?.transactions ?? [],