+++ src/UI/SignTransactionsModals/components/confirmationScreen.types.ts
@@ -2,6 +2,7 @@ import { SignPropsType } from '../types/signTransactionsModals.types';
 
 export interface ExtraConfirmationScreenPropsType {
   Screen?: (signProps: SignPropsType) => JSX.Element;
+  GuardianScreen?: SignPropsType['GuardianScreen'];
   verifyReceiverScam?: boolean;
   className?: string;
 }