+++ src/UI/SignTransactionsModals/types/signTransactionsModals.types.ts
@@ -1,6 +1,6 @@
 import { Transaction } from '@multiversx/sdk-core';
 
-import { LoginMethodsEnum } from 'types';
+import { LoginMethodsEnum, SignModalPropsType } from 'types';
 
 export interface SignPropsType {
   handleClose: () => void;
@@ -11,6 +11,8 @@ export interface SignPropsType {
   callbackRoute: string;
   className?: string;
   verifyReceiverScam?: boolean;
+  isGuarded?: boolean;
+  GuardianScreen?: SignModalPropsType['GuardianScreen'];
 }
 
 export type ScreenType = (signProps: SignPropsType) => JSX.Element;