+++ src/hooks/transactions/batch/useGetBatches.ts
@@ -1,6 +1,6 @@
+import { useMemo } from 'react';
 import { useSelector } from 'reduxStore/DappProviderContext';
 import { batchesSelector } from 'reduxStore/selectors/batchTransactionsSelectors';
-import { useMemo } from 'react';
 
 export function useGetBatches() {
   const batches = useSelector(batchesSelector);